+++ src/pyunrealsdk/unreal_bindings/bound_function.cpp
@@ -53,7 +53,8 @@ std::pair<UProperty*, std::vector<UProperty*>> fill_py_params(WrappedStruct& par
 
         // If we still have positional args left
         if (arg_idx != args.size()) {
-            py_setattr(prop, reinterpret_cast<uintptr_t>(params.base.get()), args[arg_idx++]);
+            py_setattr_direct(prop, reinterpret_cast<uintptr_t>(params.base.get()),
+                              args[arg_idx++]);
 
             if (kwargs.contains(prop->Name)) {
                 throw py::type_error(unrealsdk::fmt::format(
@@ -67,8 +68,8 @@ std::pair<UProperty*, std::vector<UProperty*>> fill_py_params(WrappedStruct& par
         if (kwargs.contains(prop->Name)) {
             // Extract the value with pop, so we can check that kwargs are empty at the
             // end
-            py_setattr(prop, reinterpret_cast<uintptr_t>(params.base.get()),
-                       kwargs.attr("pop")(prop->Name));
+            py_setattr_direct(prop, reinterpret_cast<uintptr_t>(params.base.get()),
+                              kwargs.attr("pop")(prop->Name));
             continue;
         }
 
@@ -179,7 +180,16 @@ void register_bound_function(py::module_& mod) {
                 if (args.size() == 1 && kwargs.empty() && py::isinstance<WrappedStruct>(args[0])) {
                     auto args_struct = py::cast<WrappedStruct>(args[0]);
                     if (args_struct.type == self.func) {
-                        self.call<void>(args_struct);
+                        {
+                            // Release the GIL to avoid a deadlock if ProcessEvent is locking.
+                            // If a hook tries to call into Python, it will be holding the process
+                            // event lock, and it will try to acquire the GIL.
+                            // If at the same time python code on a different thread tries to call
+                            // an unreal function, it would be holding the GIL, and trying to
+                            // acquire the process event lock.
+                            const py::gil_scoped_release gil{};
+                            self.call<void>(args_struct);
+                        }
                         return get_py_return(args_struct);
                     }
                 }
@@ -187,7 +197,10 @@ void register_bound_function(py::module_& mod) {
                 WrappedStruct params{self.func};
                 auto [return_param, out_params] = fill_py_params(params, args, kwargs);
 
-                self.call<void>(params);
+                {
+                    const py::gil_scoped_release gil{};
+                    self.call<void>(params);
+                }
 
                 return get_py_return(params, return_param, out_params);
             },